requirements.txt
pydantic==2.11.4
python-telegram-bot==22.0
requests==2.32.3
